Split workspace runtime projection helpers
