+++ src/agentic_workspace/workspace_runtime_primitives.py
@@ -144,6 +144,12 @@ from agentic_workspace.workspace_output import (
     _emit_report_text,
     _emit_setup_text,
 )
+from agentic_workspace.workspace_runtime_projection import (
+    _authority_boundary_payload,
+    _compact_authority_boundary,
+    _compact_authority_text,
+    _tiny_action_effect,
+)
 
 _GENERATED_CLI_MODULES = {
     "agentic-workspace": ("agentic_workspace._generated_cli_package_impl.cli", "generated.workspace.python.cli"),
@@ -12508,48 +12514,6 @@ def _workflow_sufficiency_payload(
     if drill_down:
         payload["drill_down"] = drill_down
     return payload
-
-
-def _authority_boundary_payload(
-    *,
-    surface: str,
-    enforced_by_aw: list[str] | None = None,
-    observed_by_aw: list[str] | None = None,
-    recommended_by_aw: list[str] | None = None,
-    candidate_routes: list[str] | None = None,
-    proof_hints: list[str] | None = None,
-    agent_owned_decisions: list[str] | None = None,
-    human_owned_decisions: list[str] | None = None,
-    rule: str | None = None,
-) -> dict[str, Any]:
-    enforced = [str(item).strip() for item in (enforced_by_aw or []) if str(item).strip()]
-    observed = [str(item).strip() for item in (observed_by_aw or []) if str(item).strip()]
-    recommended = [str(item).strip() for item in (recommended_by_aw or []) if str(item).strip()]
-    routes = [str(item).strip() for item in (candidate_routes or []) if str(item).strip()]
-    hints = [str(item).strip() for item in (proof_hints or []) if str(item).strip()]
-    agent = [str(item).strip() for item in (agent_owned_decisions or []) if str(item).strip()]
-    human = [str(item).strip() for item in (human_owned_decisions or []) if str(item).strip()]
-    if enforced:
-        authority_class = "hard-gate"
-    elif recommended or routes or hints:
-        authority_class = "advisory-support"
-    elif observed:
-        authority_class = "observed-facts"
-    else:
-        authority_class = "agent-owned"
-    return {
-        "kind": "agentic-workspace/authority-boundary/v1",
-        "surface": surface,
-        "authority_class": authority_class,
-        "enforced_by_aw": enforced,
-        "observed_by_aw": observed,
-        "recommended_by_aw": recommended,
-        "candidate_routes": routes,
-        "proof_hints": hints,
-        "agent_owned_decisions": agent,
-        "human_owned_decisions": human,
-        "reporting_rule": rule or "Report AW facts, constraints, and suggestions separately from the agent's semantic decision.",
-    }
 
 
 def _compact_continuation_state_contract(*, cli_invoke: str = DEFAULT_CLI_INVOKE) -> dict[str, Any]:
@@ -29119,23 +29083,6 @@ def _tiny_generated_surface_trust(value: dict[str, Any]) -> dict[str, Any]:
     }
 
 
-def _tiny_action_effect(value: Any, *, include_allowed: bool = True, include_resolution_commands: bool = True) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-    keys = [
-        "force",
-        "blocked_until_reconciled",
-        "resolution_selector",
-        "resolution_command",
-    ]
-    if include_resolution_commands:
-        keys.append("resolution_commands")
-    compact = {key: value[key] for key in keys if key in value}
-    if include_allowed and "allowed_now" in value:
-        compact["allowed_now"] = value["allowed_now"]
-    return compact
-
-
 def _compact_selector_next_safe_action(packet: dict[str, Any]) -> dict[str, Any]:
     compact = dict(packet)
     why = str(compact.get("why", "") or "")
@@ -29298,43 +29245,6 @@ def _tiny_workflow_sufficiency(value: Any) -> dict[str, Any]:
     compact.setdefault("sufficiency_result", value.get("decision"))
     compact["authority_boundary"] = _compact_authority_boundary(value.get("authority_boundary"))
     return {key: item for key, item in compact.items() if item not in (None, "", {})}
-
-
-def _compact_authority_boundary(value: Any) -> dict[str, Any]:
-    if not isinstance(value, dict):
-        return {}
-
-    def compact_items(key: str, limit: int) -> list[str]:
-        return [_compact_authority_text(str(item)) for item in _list_payload(value.get(key))[:limit]]
-
-    compact = {
-        "kind": value.get("kind"),
-        "surface": value.get("surface"),
-        "authority_class": value.get("authority_class"),
-        "enforced_by_aw": compact_items("enforced_by_aw", 2),
-        "observed_by_aw": compact_items("observed_by_aw", 1),
-        "recommended_by_aw": compact_items("recommended_by_aw", 1),
-        "candidate_routes": compact_items("candidate_routes", 1),
-        "agent_owned_decisions": compact_items("agent_owned_decisions", 1),
-    }
-    return {key: item for key, item in compact.items() if item not in (None, "", [])}
-
-
-def _compact_authority_text(text: str) -> str:
-    mapping = {
-        "whether delegation improves quality or cost without lowering proof": "delegation fit without lowering proof",
-        "semantic fit of candidate route to the user's task": "candidate-route semantic fit",
-        "whether to stay local when advisory delegation is not followed": "stay-local judgment",
-        "semantic work shape": "semantic work shape",
-        "proof proportionality": "proof proportionality",
-        "whether Planning is needed when no hard blocker applies": "planning need when no hard blocker applies",
-        "issue intent or acceptance when task scope is externally owned": "external issue intent or acceptance",
-    }
-    if text in mapping:
-        return mapping[text]
-    if len(text) <= 80:
-        return text
-    return text[:77].rstrip() + "..."
 
 
 def _acceptance_reconciliation_prompt_payload(*, task_text: str | None, acceptance: dict[str, Any] | None = None) -> dict[str, Any]:
